Frame genaration on amd/requirements.txt
torch
torchvision
pillow
tqdm
onnx
onnxruntime